moved struct declaration into respective file for proper organization
